refactor: move initialize method to vw_base
