+++ annif/backend/vw_base.py
@@ -1,9 +1,32 @@
 """Base class for Vowpal Wabbit based Annif backends"""
 
+import os
+from vowpalwabbit import pyvw
+from annif.exception import NotInitializedException
 from . import backend
 
 
 class VWBaseBackend(backend.AnnifLearningBackend):
     """Base class for Vowpal Wabbit based Annif backends"""
 
-    pass
+    MODEL_FILE = 'vw-model'
+    TRAIN_FILE = 'vw-train.txt'
+
+    # defaults for uninitialized instances
+    _model = None
+
+    def initialize(self):
+        if self._model is None:
+            path = os.path.join(self.datadir, self.MODEL_FILE)
+            if not os.path.exists(path):
+                raise NotInitializedException(
+                    'model {} not found'.format(path),
+                    backend_id=self.backend_id)
+            self.debug('loading VW model from {}'.format(path))
+            params = self._create_params({'i': path, 'quiet': True})
+            if 'passes' in params:
+                # don't confuse the model with passes
+                del params['passes']
+            self.debug("model parameters: {}".format(params))
+            self._model = pyvw.vw(**params)
+            self.debug('loaded model {}'.format(str(self._model)))
