+++ annif/backend/vw_ensemble.py
@@ -7,7 +7,7 @@ import annif.util
 from vowpalwabbit import pyvw
 import numpy as np
 from annif.suggestion import VectorSuggestionResult
-from annif.exception import ConfigurationException, NotInitializedException
+from annif.exception import ConfigurationException
 from . import vw_base
 from . import ensemble
 
@@ -32,28 +32,6 @@ class VWEnsembleBackend(
         'l2': (float, None),
         'passes': (int, None)
     }
-
-    MODEL_FILE = 'vw-model'
-    TRAIN_FILE = 'vw-train.txt'
-
-    # defaults for uninitialized instances
-    _model = None
-
-    def initialize(self):
-        if self._model is None:
-            path = os.path.join(self.datadir, self.MODEL_FILE)
-            if not os.path.exists(path):
-                raise NotInitializedException(
-                    'model {} not found'.format(path),
-                    backend_id=self.backend_id)
-            self.debug('loading VW model from {}'.format(path))
-            params = self._create_params({'i': path, 'quiet': True})
-            if 'passes' in params:
-                # don't confuse the model with passes
-                del params['passes']
-            self.debug("model parameters: {}".format(params))
-            self._model = pyvw.vw(**params)
-            self.debug('loaded model {}'.format(str(self._model)))
 
     @staticmethod
     def _write_train_file(examples, filename):
