+++ annif/backend/vw_multi.py
@@ -7,7 +7,7 @@ import annif.util
 from vowpalwabbit import pyvw
 import numpy as np
 from annif.suggestion import ListSuggestionResult, VectorSuggestionResult
-from annif.exception import ConfigurationException, NotInitializedException
+from annif.exception import ConfigurationException
 from . import vw_base
 from . import mixins
 
@@ -36,28 +36,6 @@ class VWMultiBackend(mixins.ChunkingBackend, vw_base.VWBaseBackend):
     SUPPORTED_ALGORITHMS = ('oaa', 'ect', 'log_multi', 'multilabel_oaa')
 
     DEFAULT_INPUTS = '_text_'
-
-    MODEL_FILE = 'vw-model'
-    TRAIN_FILE = 'vw-train.txt'
-
-    # defaults for uninitialized instances
-    _model = None
-
-    def initialize(self):
-        if self._model is None:
-            path = os.path.join(self.datadir, self.MODEL_FILE)
-            if not os.path.exists(path):
-                raise NotInitializedException(
-                    'model {} not found'.format(path),
-                    backend_id=self.backend_id)
-            self.debug('loading VW model from {}'.format(path))
-            params = self._create_params({'i': path, 'quiet': True})
-            if 'passes' in params:
-                # don't confuse the model with passes
-                del params['passes']
-            self.debug("model parameters: {}".format(params))
-            self._model = pyvw.vw(**params)
-            self.debug('loaded model {}'.format(str(self._model)))
 
     @property
     def algorithm(self):
